refactor: structure via utils
